game.d: Added invalidtile-texture as texid 0. To always be hardcoded. renderer.d: Added automagic file, line to glError-errors. Added intializations of more ogl things. Found the fukken problem that's been haunting me forever; glVertexAttribPointer _ALWAYS_ converts stuff to floats, even if we store stuff in an int shader-side; effect being like *((float**)&targetInt) = (float)value; so one needs to use a special fukken fuction for this. RE TAR DED! texture.d: Implementet texture atlas. Populate & upload. vbomaker.d: Fixed generation of horizontal surfaces. Fixed 2*offset-bug. Now to make it mixin-able, so we wont have a thousand copies of generate. main: conditional loading and unloading of libraries, depending on if we want to be a client etc. pos.d: Made mixin for toString, fuck yeah! :)

+++ gameroot/shaders/renderGR.vert
@@ -1,18 +1,22 @@
 //#version 150 core
 
 uniform mat4 VP;
-uniform vec3 offset; 
+uniform ivec3 offset; 
 
-in vec3 position;
+in ivec3 position;
 in vec2 texcoord;
-in int type;
+in uint type;
 
 out vec2 tex_texcoord;
-flat out int texId;
+flat out uint texId;
    
 void main(){
    tex_texcoord = texcoord;
    texId = type;
+
+
+
+
    gl_Position = VP * vec4(position+offset, 1.0);
 }
